requirements.txt
Flask
line-bot-sdk
requests
beautifulsoup4